+++ website/tailwind.config.js
@@ -28,5 +28,5 @@ module.exports = {
     variants: {
         extend: {},
     },
-    plugins: [],
+    plugins: [require("tailwindcss-animate")],
 }